Migrate to a Filesystem journal
